+++ src/cli/find-config.ts
@@ -1,19 +1,42 @@
 import type { CompilerSystem, Diagnostic } from '../declarations';
 import { isString, normalizePath, buildError } from '@utils';
 
-export const findConfig = async (opts: { sys: CompilerSystem; configPath: string }) => {
+/**
+ * An object containing the {@link CompilerSystem} used to find the configuration file, as well as the location on disk
+ * to search for a Stencil configuration
+ */
+export type FindConfigOptions = {
+  sys: CompilerSystem;
+  configPath: string;
+};
+
+/**
+ * The results of attempting to find a Stencil configuration file on disk
+ */
+export type FindConfigResults = {
+  configPath: string;
+  rootDir: string;
+  diagnostics: Diagnostic[];
+};
+
+/**
+ * Attempt to find a Stencil configuration file on the file system
+ * @param opts the options needed to find the configuration file
+ * @returns the results of attempting to find a configuration file on disk
+ */
+export const findConfig = async (opts: FindConfigOptions): Promise<FindConfigResults> => {
   const sys = opts.sys;
   const cwd = sys.getCurrentDirectory();
-  const results = {
+  const results: FindConfigResults = {
     configPath: null as string,
     rootDir: normalizePath(cwd),
-    diagnostics: [] as Diagnostic[],
+    diagnostics: [],
   };
   let configPath = opts.configPath;
 
   if (isString(configPath)) {
     if (!sys.platformPath.isAbsolute(configPath)) {
-      // passed in a custom stencil config location
+      // passed in a custom stencil config location,
       // but it's relative, so prefix the cwd
       configPath = normalizePath(sys.platformPath.join(cwd, configPath));
     } else {