+++ src/compiler/bundle/ext-transforms-plugin.ts
@@ -7,8 +7,18 @@ import { parseImportPath } from '../transformers/stencil-import-path';
 import type { Plugin } from 'rollup';
 import { runPluginTransformsEsmImports } from '../plugin/plugin';
 
+/**
+ * A Rollup plugin which bundles up some transformation of CSS imports as well
+ * as writing some files to disk for the `DIST_COLLECTION` output target.
+ *
+ * @param config a user-supplied configuration
+ * @param compilerCtx the current compiler context
+ * @param buildCtx the current build context
+ * @param bundleOpts bundle options for Rollup
+ * @returns a Rollup plugin which carries out the necessary work
+ */
 export const extTransformsPlugin = (
-  config: d.Config,
+  config: d.ValidatedConfig,
   compilerCtx: d.CompilerCtx,
   buildCtx: d.BuildCtx,
   bundleOpts: BundleOptions
@@ -16,14 +26,31 @@ export const extTransformsPlugin = (
   return {
     name: 'extTransformsPlugin',
 
+    /**
+     * A custom function targeting the `transform` build hook in Rollup. See here for details:
+     * https://rollupjs.org/guide/en/#transform
+     *
+     * Here we are ignoring the first argument (which contains the module's source code) and
+     * only looking at the `id` argument. We use that `id` to get information about the module
+     * in question from disk ourselves so that we can then do some transformations on it.
+     *
+     * @param _ an unused parameter (normally the code for a given module)
+     * @param id the id of a module
+     * @returns metadata for Rollup or null if no transformation should be done
+     */
     async transform(_, id) {
       if (/\0/.test(id)) {
         return null;
       }
 
+      // The `id` here was possibly previously updated using
+      // `serializeImportPath` to annotate the filepath with various metadata
+      // serialized to query-params. If that was done for this particular `id`
+      // then the `data` prop will not be null.
       const { data } = parseImportPath(id);
+
       if (data != null) {
-        let cmp: d.ComponentCompilerMeta;
+        let cmp: d.ComponentCompilerMeta | undefined = undefined;
         const filePath = normalizeFsPath(id);
         const code = await compilerCtx.fs.readFile(filePath);
         if (typeof code !== 'string') {
@@ -31,7 +58,24 @@ export const extTransformsPlugin = (
         }
 
         const pluginTransforms = await runPluginTransformsEsmImports(config, compilerCtx, buildCtx, code, filePath);
-        const commentOriginalSelector = bundleOpts.platform === 'hydrate' && data.encapsulation === 'shadow';
+
+        // We need to check whether the current build is a dev-mode watch build w/ HMR enabled in
+        // order to know how we'll want to set `commentOriginalSelector` (below). If we are doing
+        // a hydrate build we need to set this to `true` because commenting-out selectors is what
+        // gives us support for scoped CSS w/ hydrated components (we don't support shadow DOM and
+        // styling via that route for them). However, we don't want to comment selectors in dev
+        // mode when using HMR in the browser, since there we _do_ support putting stylesheets into
+        // the shadow DOM and commenting out e.g. the `:host` selector in those stylesheets will
+        // break components' CSS when an HMR update is sent to the browser.
+        //
+        // See https://github.com/ionic-team/stencil/issues/3461 for details
+        const isDevWatchHMRBuild =
+          config.flags.watch &&
+          config.flags.dev &&
+          config.flags.serve &&
+          (config.devServer?.reloadStrategy ?? null) === 'hmr';
+        const commentOriginalSelector =
+          bundleOpts.platform === 'hydrate' && data.encapsulation === 'shadow' && !isDevWatchHMRBuild;
 
         if (data.tag) {
           cmp = buildCtx.components.find((c) => c.tagName === data.tag);
@@ -39,9 +83,11 @@ export const extTransformsPlugin = (
 
           if (moduleFile) {
             const collectionDirs = config.outputTargets.filter(isOutputTargetDistCollection);
-
             const relPath = relative(config.srcDir, pluginTransforms.id);
 
+            // If we found a `moduleFile` in the module map above then we
+            // should write the transformed CSS file (found in the return value
+            // of `runPluginTransformsEsmImports`, above) to disk.
             await Promise.all(
               collectionDirs.map(async (outputTarget) => {
                 const collectionPath = join(outputTarget.collectionDir, relPath);